+++ tests/src/test/java/ca/bc/gov/open/jag/efiling/page/BceidAuthenticationPageImpl.java
@@ -1,13 +1,9 @@
 package ca.bc.gov.open.jag.efiling.page;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.Select;
 import org.springframework.beans.factory.annotation.Value;
-
-import java.io.File;
 
 public class BceidAuthenticationPageImpl extends BasePage implements AuthenticationPage {
 
@@ -19,10 +15,6 @@ public class BceidAuthenticationPageImpl extends BasePage implements Authenticat
 
     @Value("${PASSWORD_BCEID}")
     private String bceidPassword;
-
-    @Value("classpath:data/test-document.pdf")
-    private File pdfDocumentToUpload;
-
 
     //Page Objects:
     @FindBy(id = "zocial-bceid")
@@ -36,33 +28,6 @@ public class BceidAuthenticationPageImpl extends BasePage implements Authenticat
 
     @FindBy(xpath = "//input[@name='btnSubmit']")
     private WebElement submitButton;
-
-    @FindBy(xpath = "//*[@data-testid='dropdownzone']/div/input")
-    private WebElement fileUploadInput;
-
-    @FindBy(xpath = "//button[@data-test-id='generate-url-btn']")
-    private WebElement filingPackageButton;
-
-    public void redirectToEfilingHub() {
-        uploadAPdfOnDemoClient();
-        selectDocumentTypeAndSubmit();
-    }
-
-    public void uploadAPdfOnDemoClient() {
-        try {
-            wait.until(ExpectedConditions.visibilityOf(fileUploadInput)).sendKeys(pdfDocumentToUpload.getPath());
-        } catch (org.openqa.selenium.TimeoutException ex) {
-            fileUploadInput.sendKeys(pdfDocumentToUpload.getPath());
-            System.out.println(pdfDocumentToUpload.getPath());
-        }
-    }
-
-    public void selectDocumentTypeAndSubmit() {
-        Select selectDocumentType = new Select(this.driver.findElement(By.id("dropdown")));
-        selectDocumentType.selectByValue("AFF");
-        filingPackageButton.click();
-        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//button[@data-test-id='continue-btn']")));
-    }
 
     public void loginWithBceid() {
         wait.until(ExpectedConditions.titleContains("Log in to Family Law Act Application"));
@@ -88,6 +53,5 @@ public class BceidAuthenticationPageImpl extends BasePage implements Authenticat
     public void signIn() {
         goTo(efilingAdminUrl);
         loginWithBceid();
-        redirectToEfilingHub();
     }
 }
