+++ tests/src/test/resources/application.properties
@@ -1,13 +1,10 @@
-# ACCESS URL
+#ACCESSURL
 EFILING_ADMIN=${EFILING_ADMIN_URL}
 
-# BCSC CREDS
+#BCSCCREDS
 BCSC_CARD_NUMBER=${BCSC_CARD_NUMBER}
 BCSC_PASSCODE=${BCSC_PASSCODE}
 
-# BCEID CREDS
+#BCEIDCREDS
 USERNAME_BCEID=${USERNAME_BCEID}
 PASSWORD_BCEID=${PASSWORD_BCEID}
-
-# TEST RUN CONFIG
-TEST=${PROVIDER}
